Removing pytest-mpi
