+++ test/test_newat.py
@@ -17,7 +17,6 @@ def triangle_area(p1, p2, p3):
 
     return abs(x1 * (y2 - y3) + x2 * (y3 - y1) + x3 * (y1 - y2)) / 2
 
-@pytest.mark.mpi_skip()
 def test_correct_receiver_location_generation2D():
     """Tests if receiver locations where generated correctly"""
     comm = spyro.utils.mpi_init(model)
@@ -28,7 +27,6 @@ def test_correct_receiver_location_generation2D():
 
     assert np.allclose(receivers, answer)
 
-@pytest.mark.mpi_skip()
 def test_correct_receiver_to_cell_location2D():
     """Tests if the receivers where located in the correct cell"""
     comm = spyro.utils.mpi_init(model)
@@ -88,7 +86,6 @@ def test_correct_receiver_to_cell_location2D():
 
     assert all([test1, test2, test3])
 
-@pytest.mark.mpi_skip()
 def test_correct_at_value2D():
     comm = spyro.utils.mpi_init(model)
     model["opts"]["degree"] = 3
@@ -120,7 +117,6 @@ def test_correct_at_value2D():
 
     assert all([test1, test2])
 
-@pytest.mark.mpi_skip()
 def test_correct_at_value2D_quad():
     model_quad = deepcopy(model)
     comm = spyro.utils.mpi_init(model_quad)
@@ -169,7 +165,6 @@ def tetrahedral_volume(p1, p2, p3, p4):
 
     return volume
 
-@pytest.mark.mpi_skip()
 def test_correct_receiver_location_generation3D():
     """Tests if receiver locations where generated correctly"""
 
@@ -184,7 +179,6 @@ def test_correct_receiver_location_generation3D():
 
     assert np.allclose(receivers.receiver_locations, answer)
 
-@pytest.mark.mpi_skip()
 def test_correct_receiver_to_cell_location3D():
     """Tests if the receivers where located in the correct cell"""
 
@@ -258,7 +252,6 @@ def test_correct_receiver_to_cell_location3D():
 
     assert all([test1, test2, test3])
 
-@pytest.mark.mpi_skip()
 def test_correct_at_value3D():
     test_model2 = deepcopy(model3D)
     test_model2["acquisition"]["num_receivers"] = 3
@@ -299,7 +292,7 @@ if __name__ == "__main__":
     test_correct_receiver_location_generation2D()
     test_correct_receiver_to_cell_location2D()
     test_correct_at_value2D()
-    test_correct_at_value2d_quad()
+    test_correct_at_value2D_quad()
     test_correct_receiver_location_generation3D()
     test_correct_receiver_to_cell_location3D()
     test_correct_at_value3D()
