+++ test/test_tools.py
@@ -27,7 +27,6 @@ def triangle_area(p1, p2, p3):
 
     return abs(x1 * (y2 - y3) + x2 * (y3 - y1) + x3 * (y1 - y2)) / 2
 
-@pytest.mark.mpi_skip()
 def test_mesh_generation_for_grid_calc():
     grid_point_calculator_parameters = {
         ## Experiment parameters
@@ -59,37 +58,7 @@ def test_mesh_generation_for_grid_calc():
         model["mesh"]["meshfile"] = "meshes/2Dhomogeneous"+str(G)+".msh"
         model = spyro.tools.generate_mesh(model, G, comm)
 
-@pytest.mark.mpi(min_size=5)
-def test_grid_calc2d():
-    grid_point_calculator_parameters = {
-        ## Experiment parameters
-        'source_frequency' : 5.0, # Here we define the frequency of the Ricker wavelet source
-        'minimum_velocity_in_the_domain' :  1.429, # The minimum velocity present in the domain.
-        # if an homogeneous test case is used this velocity will be defined in the whole domain.
-        'velocity_profile_type': 'homogeneous', # Either or heterogeneous. If heterogeneous is 
-        #chosen be careful to have the desired velocity model below.
-        'velocity_model_file_name': 'vel_z6.25m_x12.5m_exact.segy',
-        'FEM_method_to_evaluate' : 'KMV', # FEM to evaluate such as `KMV` or `spectral` (GLL nodes on quads and hexas)
-        'dimension' : 2, # Domain dimension. Either 2 or 3.
-        'receiver_setup' : 'near', #Either near or line. Near defines a receiver grid near to the source,
-        # line defines a line of point receivers with pre-established near and far offsets.
 
-        ## Line search parameters
-        'reference_degree': 4, # Degree to use in the reference case (int)
-        'G_reference': 8.0, # grid point density to use in the reference case (float)
-        'desired_degree': 4, # degree we are calculating G for. (int)
-        'G_initial': 7.0, # Initial G for line search (float)
-        'accepted_error_threshold': 0.07, 
-        'g_accuracy': 1e-1
-        }
-
-
-    G = spyro.tools.minimum_grid_point_calculator(grid_point_calculator_parameters)
-    inside = (6.9< G and G<8.0)
-    print(G)
-    assert inside
-
-@pytest.mark.mpi_skip()
 def test_input_models_receivers():
     test1 = True #testing if 2D receivers are inside the domain on an homogeneous case
     grid_point_calculator_parameters = {
@@ -198,7 +167,6 @@ def test_input_models_receivers():
 
     assert all([test1, test2])
 
-@pytest.mark.mpi_skip()
 def test_input_models_receivers_heterogeneous():
     test1 = True #testing if 2D receivers bins are inside the domain on an heterogeneous case
     grid_point_calculator_parameters = {
@@ -305,4 +273,3 @@ def test_input_models_receivers_heterogeneous():
 
 if __name__ == "__main__":
     test_mesh_generation_for_grid_calc()
-    test_grid_calc2d()
